add mapeamentos

+++ ranking-core/src/main/java/com/mhj/ranking/entity/Classificacao.java
@@ -34,7 +34,7 @@ public class Classificacao implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @JoinColumn(name = "id_equipe", referencedColumnName = "id")
     @ManyToOne(optional = false, fetch = FetchType.LAZY)
     private Equipe idEquipe;
@@ -48,15 +48,15 @@ public class Classificacao implements Serializable {
     public Classificacao() {
     }
 
-    public Classificacao(Integer id) {
+    public Classificacao(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
